requirements.txt
streamlit>=1.39
pymupdf>=1.24.9
faiss-cpu>=1.8.0
sentence-transformers>=3.2.1
pandas>=2.2.2
numpy>=1.26
rapidfuzz>=3.9.7
python-dotenv>=1.0.1
google-generativeai>=0.7
python-docx>=1.1.2
scikit-learn>=1.5.2